+++ tests/Neo.L2.Abstractions.UnitTests/UT_NeoHubDeployReport.cs
@@ -218,6 +218,7 @@ public class UT_NeoHubDeployReport
                 + "IsBatcherCheckpointAligned / "
                 + "IsMetricsHttpHealthy / MetricsHttpHealthFailures / "
                 + "IsLocalHostHealthy / LocalHostHealthFailures / IsLocalHostHealthyAsync / "
+                + "IsPipelineHealthyAsync / GetPipelineHealthFailuresAsync / "
                 + "StartMetricsHttp(/readyz defaults to IsOfflinePassportComplete) / "
                 + "HasExpectedNetwork / HasScannerDeployHeights / IsOfflinePassportComplete / "
                 + "OfflinePassportFailures / "